chore: format test files
