+++ tests/solidity/suites/precompiles/test/bech32/methods.js
@@ -1,4 +1,4 @@
-const { expect } = require('chai');
+const {expect} = require('chai');
 const hre = require('hardhat');
 
 describe('Bech32', function () {
